Use text template
A bit unsafe hope we escape them all.

reason is thml template is too slow. Well I ran it on a very weak
machine and exprience long delay in the first response. The next is fine
(cached?) but after a while it slows down again, due to the box has very
limited memory forcing go to clear cached in GC?